Rearrange spec_builders
