+++ pysetup/spec_builders/__init__.py
@@ -1,0 +1,16 @@
+from .phase0 import Phase0SpecBuilder
+from .altair import AltairSpecBuilder
+from .bellatrix import BellatrixSpecBuilder
+from .capella import CapellaSpecBuilder
+from .deneb import DenebSpecBuilder
+from .eip6110 import EIP6110SpecBuilder
+from .whisk import WhiskSpecBuilder
+
+
+spec_builders = {
+    builder.fork: builder
+    for builder in (
+        Phase0SpecBuilder, AltairSpecBuilder, BellatrixSpecBuilder, CapellaSpecBuilder, DenebSpecBuilder,
+        EIP6110SpecBuilder, WhiskSpecBuilder,
+    )
+}
